Commit progress before adjusting Table members
Having header items in the same vector as actual items is proving a
problem for sorting, they were initially merged to reduce the number of
parameters passed to certain functions, specifically column creation
functions.

Instead the header items will now form an an array of two Items.

+++ src/cat/item_sorter.h
@@ -3,11 +3,11 @@
 template<typename T>
 class Compare {
 	protected:
-		std::vector<int> _elements;
+		std::vector<size_t> _elements;
 
 		auto _element(const T &t, const int element);
 	public:
-		Compare(std::vector<int> &&elements);
+		Compare(std::vector<size_t> &&elements);
 
 		bool operator()(const T &a, const T &b);
 };
